+++ src/redturtle/volto/events.py
@@ -1,3 +1,7 @@
+from plone import api
+from redturtle.volto.adapters.namechooser import NormalizingNameChooser
+
+
 import os
 
 
@@ -16,3 +20,8 @@ def manage_auth_token(event):
         auth_token = request.cookies.get("auth_token")
         if auth_token:
             request._auth = "Bearer " + auth_token
+
+
+def avoid_spaces_in_shortname(obj, event):
+    fixed_id = NormalizingNameChooser(obj).chooseName(obj.id, obj)
+    api.content.rename(obj, new_id=fixed_id)