+++ src/redturtle/volto/tests/test_spaces_in_id.py
@@ -63,4 +63,4 @@ class TestCreation(unittest.TestCase):
             },
         )
 
-        self.assertEqual(response.status_code, 201)
+        self.assertEqual(response.json()["id"], "aabbu")