+++ src/redturtle/volto/tests/test_blocks_linkintegrity.py
@@ -10,7 +10,6 @@ from zope.lifecycleevent import ObjectModifiedEvent
 
 import unittest
 
-
 HAS_PLONE_6 = getattr(
     import_module("Products.CMFPlone.factory"), "PLONE60MARKER", False
 )